wip: parte de cadastrar funcionando, resto ainda precisa fazer

+++ src/main/java/buteco/controller/produtos/dto/CadProdutosResponse.java
@@ -3,7 +3,7 @@ package buteco.controller.produtos.dto;
 public record CadProdutosResponse(
         String nome,
         double preco_venda,
-        int fk_id_categoria,
-        int fk_id_grupo,
+        Long fk_id_categoria,
+        Long fk_id_grupo,
         String obsevacao){
 }
